konia-react-migration/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pymongo==4.7.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
dnspython==2.6.1
pandas==2.2.2
plotly==5.22.0
numpy==1.26.4
requests==2.31.0
pytest==8.2.0
httpx==0.27.0
